add package:synchronized to prevent file corruption (#38)
